refactor(file-service, dict-service): 优化代码细节
- 在 FilePartServiceImpl 中，将 count 方法的返回类型转换为 int
- 在 SysDictTypeServiceImpl 中：
  - 将 count 方法的返回类型转换为 int
  - 优化异常消息格式
  - 使用常量替代硬编码的字符串

+++ common-service-module-all/file-service-module/src/main/java/com/minimalism/file/service/impl/FilePartServiceImpl.java
@@ -142,7 +142,7 @@ public class FilePartServiceImpl extends ServiceImpl<FilePartMapper, FilePart> i
                 .eq(ObjectUtils.isNotEmpty(fileId),FilePart::getFileId, fileId)
                 .eq(StrUtil.isNotBlank(identifier),FilePart::getPartCode, identifier)
                 .groupBy(FilePart::getPartCode);
-        return count(wrapper);
+        return (int) count(wrapper);
     }
 
     @Override
